requirements.txt

protobuf==3.14.0
six==1.15.0
grpcio==1.32.0
grpcio-tools==1.32.0
